# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.CreateTransactionRequest:
    properties:
      barang_id:
        example: 1
        type: integer
    required:
    - barang_id
    type: object
  domain.ItemCategory:
    enum:
    - Buku
    - Elektronik
    - Perabotan
    - Kos-kosan
    - Lainnya
    type: string
    x-enum-varnames:
    - CategoryBuku
    - CategoryElektronik
    - CategoryPerabotan
    - CategoryKosKosan
    - CategoryLainnya
  domain.ItemResponse:
    properties:
      created_at:
        type: string
      deskripsi:
        type: string
      gambar:
        type: string
      harga:
        type: number
      id:
        type: integer
      kategori:
        $ref: '#/definitions/domain.ItemCategory'
      nama_barang:
        type: string
      penjual:
        $ref: '#/definitions/domain.UserResponse'
      penjual_id:
        type: integer
      status:
        $ref: '#/definitions/domain.ItemStatus'
    type: object
...
  domain.ItemStatus:
    enum:
    - Tersedia
    - Terjual
    - Dihapus
    type: string
    x-enum-varnames:
    - StatusTersedia
    - StatusTerjual
    - StatusDihapus
...
  domain.Role:
    enum:
    - user
    - admin
    type: string
    x-enum-varnames:
    - RoleUser
    - RoleAdmin
...
  domain.TransactionResponse:
    properties:
      barang:
        $ref: '#/definitions/domain.ItemResponse'
      barang_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      pembeli:
        $ref: '#/definitions/domain.UserResponse'
      pembeli_id:
        type: integer
      status_transaksi:
        $ref: '#/definitions/domain.TransactionStatus'
      tanggal_transaksi:
        type: string
    type: object
  domain.TransactionStatus:
    enum:
    - Pending
    - Selesai
    - Dibatalkan
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusSelesai
    - StatusDibatalkan
...
  domain.UpdateTransactionStatusRequest:
    properties:
      status:
        allOf:
        - $ref: '#/definitions/domain.TransactionStatus'
        enum:
        - Pending
        - Selesai
        - Dibatalkan
        example: Selesai
    required:
    - status
    type: object
...
  domain.UserResponse:
    properties:
      alamat:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      nama:
        type: string
      no_hp:
        type: string
      role:
        $ref: '#/definitions/domain.Role'
      updated_at:
        type: string
    type: object
...
  utils.Meta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  utils.PaginatedResponse:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/utils.Meta'
      success:
        type: boolean
    type: object
  utils.StandardResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@jubel.app
    name: Jubel API Support
  description: API untuk aplikasi jual beli barang bekas Jubel
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Jubel API
  version: "1.0"
paths:
...
  /transactions:
    get:
      consumes:
      - application/json
      description: Mendapatkan daftar semua transaksi (admin only)
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.TransactionResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardResponse'
      security:
      - BearerAuth: []
      summary: List all transactions
      tags:
      - transactions
    post:
      consumes:
      - application/json
      description: Membuat transaksi baru untuk pembelian barang
      parameters:
      - description: Transaction data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateTransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.TransactionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardResponse'
      security:
      - BearerAuth: []
      summary: Create a transaction
      tags:
      - transactions
  /transactions/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus transaksi berdasarkan ID
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardResponse'
      security:
      - BearerAuth: []
      summary: Delete transaction
      tags:
      - transactions
    get:
      consumes:
      - application/json
      description: Mendapatkan detail transaksi berdasarkan ID
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.TransactionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.StandardResponse'
      security:
      - BearerAuth: []
      summary: Get transaction by ID
      tags:
      - transactions
  /transactions/{id}/status:
    patch:
      consumes:
      - application/json
      description: Memperbarui status transaksi (Pending, Selesai, Dibatalkan)
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      - description: Transaction status data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateTransactionStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardResponse'
      security:
      - BearerAuth: []
      summary: Update transaction status
      tags:
      - transactions
  /transactions/as-pembeli:
    get:
      consumes:
      - application/json
      description: Mendapatkan daftar transaksi dimana user adalah pembeli
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.TransactionResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardResponse'
      security:
      - BearerAuth: []
      summary: Get my transactions as buyer
      tags:
      - transactions
  /transactions/as-penjual:
    get:
      consumes:
      - application/json
      description: Mendapatkan daftar transaksi dimana user adalah penjual
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.TransactionResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardResponse'
      security:
      - BearerAuth: []
      summary: Get my transactions as seller
      tags:
      - transactions
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"